requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
Authlib
PyJWT
bcrypt
mysqlclient 